Task 5 complete, added modules and re-written fillComments function
