+++ js/data.js
@@ -1,0 +1,42 @@
+const NAMES = [
+  'Александр',
+  'Иван',
+  'Максим',
+  'Егор',
+  'Михаил',
+  'Андрей',
+  'Марк',
+  'Сергей',
+  'Роман',
+  'Владимир',
+  'Матвей',
+  'Денис',
+  'Павел',
+  'Степан',
+  'Глеб',
+  'Григорий',
+  'Руслан',
+  'Кирилл',
+  'Вадим',
+  'Константин',
+  'Олег',
+  'Ярослав',
+  'Антон',
+  'Виктор',
+  'Артём'
+];
+const COMMENTS = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'];
+const DESCRIPTIONS = [
+  'Смотрите какая картинка',
+  'Моя лучшая работа!',
+  'Я хочу се....анс у психиатра',
+  'смотрите какую фотку забацал'
+];
+
+export {NAMES, COMMENTS, DESCRIPTIONS};
